+++ special-pages/pages/history/app/constants.js
@@ -3,4 +3,3 @@ export const OVERSCAN_AMOUNT = 5;
 export const BTN_ACTION_ENTRIES_MENU = 'entries_menu';
 export const BTN_ACTION_DELETE_RANGE = 'deleteRange';
 export const KNOWN_ACTIONS = /** @type {const} */ ([BTN_ACTION_ENTRIES_MENU, BTN_ACTION_DELETE_RANGE]);
-export const EVENT_RANGE_CHANGE = 'range-change';
