+++ special-pages/pages/history/app/global/Providers/QueryProvider.js
@@ -1,8 +1,6 @@
 import { createContext, h } from 'preact';
 import { useCallback, useContext } from 'preact/hooks';
-import { signal, useComputed, useSignal, useSignalEffect } from '@preact/signals';
-import { useHistoryServiceDispatch } from './HistoryServiceProvider.js';
-import { useSettings } from '../../types.js';
+import { signal, useSignal } from '@preact/signals';
 
 /**
  * @typedef {import('../../../types/history.ts').Range} Range
@@ -99,138 +97,4 @@ export function useQueryContext() {
  */
 export function useQueryDispatch() {
     return useContext(QueryDispatch);
-}
-
-/**
- * Synchronizes the `derivedRange` signal with the browser's URL and issues a
- * 'search-commit'. This allows any part of the application to change the range and
- * have it synchronised to the URL + trigger a search
- */
-export function useSearchCommitForRange() {
-    const dispatch = useHistoryServiceDispatch();
-    const query = useQueryContext();
-    const derivedRange = useComputed(() => {
-        return query.value.range;
-    });
-
-    useSignalEffect(() => {
-        let timer;
-        let counter = 0;
-        const sub = derivedRange.subscribe((nextRange) => {
-            if (counter === 0) {
-                counter += 1;
-                return;
-            }
-            const url = new URL(window.location.href);
-
-            url.searchParams.delete('q');
-            url.searchParams.delete('range');
-
-            if (nextRange !== null) {
-                url.searchParams.set('range', nextRange);
-                window.history.replaceState(null, '', url.toString());
-                dispatch({ kind: 'search-commit', params: new URLSearchParams(url.searchParams) });
-            }
-        });
-
-        return () => {
-            sub();
-            clearTimeout(timer);
-        };
-    });
-}
-
-/**
- * Updates the URL with the latest search term (if present) and dispatches a custom event with the updated query parameters.
- * Debounces the updates based on the `settings.typingDebounce` value to avoid frequent URL state changes during typing.
- *
- * This hook uses a signal effect to listen for changes in the `derivedTerm` and updates the browser's URL accordingly, with debounce support.
- * It dispatches an `EVENT_SEARCH_COMMIT` event to notify other components or parts of the application about the updated search parameters.
- *
- */
-export function useURLReflection() {
-    const settings = useSettings();
-    const query = useQueryContext();
-    useSignalEffect(() => {
-        let timer;
-        let count = 0;
-        const unsubscribe = query.subscribe((nextValue) => {
-            if (count === 0) return (count += 1);
-            clearTimeout(timer);
-            if (nextValue.term !== null) {
-                const term = nextValue.term;
-                timer = setTimeout(() => {
-                    const url = new URL(window.location.href);
-
-                    url.searchParams.set('q', term);
-                    url.searchParams.delete('range');
-                    url.searchParams.delete('domain');
-
-                    if (term.trim() === '') {
-                        url.searchParams.delete('q');
-                    }
-
-                    window.history.replaceState(null, '', url.toString());
-                }, settings.urlDebounce);
-            }
-            if (nextValue.domain !== null) {
-                const url = new URL(window.location.href);
-                url.searchParams.set('domain', nextValue.domain);
-                url.searchParams.delete('q');
-                url.searchParams.delete('range');
-
-                window.history.replaceState(null, '', url.toString());
-            }
-            return null;
-        });
-
-        return () => {
-            unsubscribe();
-            clearTimeout(timer);
-        };
-    });
-}
-
-/**
- * Updates the URL with the latest search term (if present) and dispatches a custom search commit event.
- * Utilizes a debounce mechanism to ensure the URL updates are not performed too often during typing.
- *
- * Workflow:
- * - Listens for changes in the `derivedTerm` signal.
- * - For the first signal emission (`counter === 0`), skips processing to avoid triggering on initial load.
- * - Debounces subsequent changes using `settings.typingDebounce`.
- * - If a non-null value is provided, constructs query parameters with the new term and dispatches
- *   an `EVENT_SEARCH_COMMIT` event with the updated parameters.
- * - If the value is `null`, no action is taken.
- */
-export function useSearchCommit() {
-    const dispatch = useHistoryServiceDispatch();
-    const settings = useSettings();
-    const query = useQueryContext();
-    useSignalEffect(() => {
-        let timer;
-        let count = 0;
-        const unsubscribe = query.subscribe((next) => {
-            if (count === 0) return (count += 1);
-            clearTimeout(timer);
-            if (next.term !== null) {
-                const term = next.term;
-                timer = setTimeout(() => {
-                    const params = new URLSearchParams();
-                    params.set('q', term);
-                    dispatch({ kind: 'search-commit', params });
-                }, settings.typingDebounce);
-            }
-            if (next.domain !== null) {
-                const params = new URLSearchParams();
-                params.set('domain', next.domain);
-                dispatch({ kind: 'search-commit', params });
-            }
-            return null;
-        });
-        return () => {
-            unsubscribe();
-            clearTimeout(timer);
-        };
-    });
 }
